chore: delete packet-decoder
This was an experimental AX.25 packet decoder using GNU radio. I've
removed it as it clutters the repository and we're not decoding AX.25
packets at the moment. We may come back to it at a later date when we
are preparing for a Day-in-the-Life test in which we will actually be
decoding RF signals from the satellite through the antenna.