Improve multiprocessing
KeyboardInterrupts are handled properly in each process. The message queues are just flushed in this case. When exiting normally, the queues will be processed until they're empty. The serial handler stops reading new messages though.

finalize_session_log() is done in the logger process itself. The query command is also commented out currently because I removed file handling from the main process. I might just go back to how it was, with the logger sending the file name to the main process, but that doesn't feel right. Maybe do_query() could send a request to the logger, and the response is piped back? Query could also just be removed potentially, it's a little outdated with the logging improvements.