Feat: Simplify Stopping of Subprocesses
Now uses a flag to stop the processes. Also removed the code that
cleared queues after the process is killed since it added complexity but
if the process is being killed it likely means the program is exiting in
which case it's not that important that we clear the queues.