Write to Serial Reader in Virtual Mode Again
Now that we don't have to wait for the queue to empty, we can still
write to the serial reader. Why do I do this? Because I think most of
the code should be blind to whether you are in virtual mode, so that the
behaviour is the same no matter what. The way we've set up the serial
reader is not very general so there's no way to consume messages, but I
can imagine creating some kind of `Port` class with `port.write` and
`port.read`, so that we can easily make a virtual port that is treated
identically to normal ports in the code. That's not for this PR though.