Refactor: Simplify Log File Writing Code
Made several simplifications. The biggest difference to the code came
from storing the log as a string and only writing to disk at the end of
the program. This avoided much of the file I/O and exception handling
code and removed the need for an init and finalize functions for the
log files. I also simplified the formatting in the file.

This solution uses strings to maintain the message log throughout
execution. However, strings are a bit heavyweight (though perhaps less
than file operations). If this approach ever gets slow, we can replace
it with a list of Message objects, and convert them to YAML at the end.