Implement cast syntax and negative values
Integer width and signedness is handled more carefully, with feedback for user error. It's important to consider that this interprets values based on the type provided (ex. (i8)0xBA is sent as 0xBA with the assumption that the user meant to give the value in two's complement - it is not rejected for overflow)