Add instructions for installing uv and soti
These instructions will need to be modified once soti is published on
PyPI.

Note that since we specify which branch to install from, I've chosen to
specify dev as the install branch for now. Once this branch is merged
into dev, the instructions will work, but they'll be broken until then.
To test, find the git+https:// URLs and replace @dev with
@feature/uv-package-manager.